changed the amount data type from double to decimal

+++ PensionContManageSystem/Domain/DTOs/ResponseDto/StatementResponseDto.cs
@@ -3,9 +3,9 @@
     public class StatementResponseDto
     {
         public int MemberId { get; set; }
-        public double TotalAmountOfVoluntaryContribution { get; set; }
-        public double TotalAmountOfMonthlyContribution { get; set; }
-        public double TotalContribution { get; set; }
+        public decimal TotalAmountOfVoluntaryContribution { get; set; }
+        public decimal TotalAmountOfMonthlyContribution { get; set; }
+        public decimal TotalContribution { get; set; }
         public List<ContributionResponseDto> Contributions { get; set; }
         public string From { get; set; }
         public string To { get; set; }
